Changed oath text with God to match without
Before, selecting to do an oath with God as opposed to without (i.e. affirmation) would mistakenly add a clause "recognizing the supreme authority of Niabar." This clause was never intended to be part of the oath and as such was removed. No citizen has taken the oath with the additional clause (if they had, it would've required delicate handling).